+++ InventoryTools/Logic/Filters/IsCollectibleFilter.cs
@@ -1,14 +1,23 @@
+using System.Collections.Generic;
+using System.Linq;
 using AllaganLib.GameSheets.Sheets.Rows;
 using CriticalCommonLib.Models;
 
 using InventoryTools.Logic.Filters.Abstract;
 using InventoryTools.Services;
+using Lumina.Excel;
+using Lumina.Excel.Sheets;
 using Microsoft.Extensions.Logging;
 
 namespace InventoryTools.Logic.Filters
 {
     public class IsCollectibleFilter : BooleanFilter
     {
+
+        public IsCollectibleFilter(ILogger<IsCollectibleFilter> logger, ImGuiService imGuiService) : base(logger, imGuiService)
+        {
+        }
+
         public override string Key { get; set; } = "Collectible";
         public override string Name { get; set; } = "Is Collectible?";
         public override string HelpText { get; set; } = "Is the item Collectible?";
@@ -34,16 +43,12 @@ namespace InventoryTools.Logic.Filters
             var currentValue = CurrentValue(configuration);
             if (currentValue == null) return true;
 
-            if(currentValue.Value && item.Base.IsCollectable)
+            if(currentValue.Value && item.IsCollectable)
             {
                 return true;
             }
 
-            return !currentValue.Value && !item.Base.IsCollectable;
-        }
-
-        public IsCollectibleFilter(ILogger<IsCollectibleFilter> logger, ImGuiService imGuiService) : base(logger, imGuiService)
-        {
+            return !currentValue.Value && !item.IsCollectable;
         }
     }
 }